Refactor ocr filter/capture to view model
